Manager code files and Dockerfile update
